[refactor](#11) character 위치 변경 (user->learning)
