+++ src/main/java/com/poweranger/hai_duo/user/application/service/UserService.java
@@ -6,12 +6,12 @@ import com.poweranger.hai_duo.learning.api.dto.LevelUpResultDto;
 import com.poweranger.hai_duo.user.api.dto.UserDto;
 import com.poweranger.hai_duo.learning.api.factory.LevelUpResultFactory;
 import com.poweranger.hai_duo.user.api.factory.UserFactory;
-import com.poweranger.hai_duo.user.application.processor.CharacterUpgradeProcessor;
+import com.poweranger.hai_duo.learning.application.processor.CharacterUpgradeProcessor;
 import com.poweranger.hai_duo.learning.application.processor.LevelUpProcessor;
-import com.poweranger.hai_duo.user.domain.entity.mysql.GameCharacter;
+import com.poweranger.hai_duo.learning.domain.entity.GameCharacter;
 import com.poweranger.hai_duo.learning.domain.entity.Level;
 import com.poweranger.hai_duo.user.domain.entity.mysql.User;
-import com.poweranger.hai_duo.user.domain.repository.CharacterRepository;
+import com.poweranger.hai_duo.learning.domain.repository.CharacterRepository;
 import com.poweranger.hai_duo.learning.domain.repository.LevelRepository;
 import com.poweranger.hai_duo.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
